V2 Functionality Complete (#10)
Added a new table in the database for  active announcements. This stops the bot's permission checking loop from needing to be added/restarted/etc. Users who are supposed to have permissions are ignored from the loop.

Removed the need for /qr_codes/.

Events are now removed from /lockdown

Secrets are now sent alongside the message for manual authy/google auth.

Co-authored-by: ElevatedAlexandra <109626174+ElevatedAlexandra@users.noreply.github.com>